+++ packages/trpc/server/routers/viewer/workflows/get.schema.ts
@@ -1,7 +1,2 @@
-import { z } from "zod";
-
-export const ZGetInputSchema = z.object({
-  id: z.number(),
-});
-
-export type TGetInputSchema = z.infer<typeof ZGetInputSchema>;
+export { ZGetInputSchema } from "@calcom/lib/server/repository/workflow";
+export type { TGetInputSchema } from "@calcom/lib/server/repository/workflow";
